Implemented miner side target deadline
Before this commit, the miner only knew about the target deadline of the pool.
Now it is possible to set a target deadline manually.
The miner will still take the lower of both target deadlines, what means, that if the pool has a lower target deadline then yours, the miner will take the pool's one.

Related to issue #38